terraform/modules/common/any/payload/linux/modules/responder-reverse-shell-multistage/resources/requirements.txt
pwncat==0.1.2
PyYAML==6.0.1
kubernetes==29.0.0
